Created slices to implement reducers
